Add files via upload
Not much to say this is just for those who want to injext metadata on VR HLS files that just might be in the hundreds .. anyway this simple exe will allow you to pick as many as you want and will the inject them all with the metadata. I spent 7 hours on this and frankly i am very pissed off on how people upload so called free software unto github and do not include a exe file in them. to me it seems it should be common sense that every Complete program / version is released with a .exe file that at least has a GUI on it. These are not words i am supposed to f******* know but the last 7 hours have been s***. Lastly the genius that created FFMPEG you are great but one lazy a** m*****f*****. How the hell do you make somehting so useful but cant be bothered to make the most simple GUI for all of it. FFMPEG sure it has gui made by other fans who give a s*** their is nothing quite like an offical release by the creators. 
ffs if this gets deleted i dont care but those who have 100% looked for a "Multi Spatial Media Metadata Injector are the only ones missing out.